feat(agents): Add shared UI utilities and refactor agent pages
- Create lib/pap-ui-utils.ts with shared types, formatters, and validation
- Deduplicate formatUptime, timeAgo, formatDate across agent pages
- Add heartbeat interval constants from pap-constants.ts
- Add agent name validation matching backend policy
- Refactor agent pages to use shared utilities
